Refactor objc2_test_utils
