--- objc2_test_utils/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "objc2_test_utils"
-version = "0.0.2"
-authors = ["Steven Sheldon", "Mads Marquart <mads@marquart.dk>"]
-edition = "2018"
-
-description = "Utilities for testing Objective-C interop."
-keywords = ["objective-c", "macos", "ios", "testing"]
-categories = [
-    "development-tools::ffi",
-    "no-std",
-    "development-tools::testing",
-    "os::macos-apis",
-]
-repository = "https://github.com/madsmtm/objc2"
-documentation = "https://docs.rs/objc2_test_utils/"
-license = "MIT"
-
-build = "build.rs"
-
-[dependencies]
-block-sys = { path = "../block-sys" }
-
-[build-dependencies]
-cc = "1.0"
