+++ tests/extern/block_utils.c
